project1/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Jinja2

python-dotenv

azure-cognitiveservices-vision-computervision
msrest

google-generativeai        # Gemini용 (2,5단계에서 사용)
diffusers                  # Stable Diffusion (3단계: 로컬 추론용)
transformers
accelerate
safetensors
Pillow

requests


torch
